+++ src/fetch.js
@@ -1,13 +1,14 @@
 import HttpError from './HttpError';
 
 const fetchJson = (url, options = {}) => {
-    const requestHeaders = options.headers || new Headers({
-        'Accept': 'application/vnd.api+json',
-    });
-    if (!(options && options.body && options.body instanceof FormData)) {
+    const requestHeaders = options.headers || new Headers();
+    if (!requestHeaders.has('Accept')){
+        requestHeaders.set('Accept', 'application/vnd.api+json')
+    }
+    if (!(options && options.body && options.body instanceof FormData) && !requestHeaders.has('Content-Type')) {
         requestHeaders.set('Content-Type', 'application/vnd.api+json');
     }
-    if (options.user && options.user.authenticated && options.user.token) {
+    if (options.user && options.user.authenticated && options.user.token && !requestHeaders.has('Authorization')) {
         requestHeaders.set('Authorization', options.user.token);
     }
     return fetch(url, { ...options, headers: requestHeaders })
@@ -35,7 +36,9 @@ export const jsonApiHttpClient = (url, options = {}) => {
     if (!options.headers) {
         options.headers = new Headers({ 'Accept': 'application/vnd.api+json' });
     }
-    options.headers.set('Content-Type', 'application/vnd.api+json');
+    if (!options.headers.has('Content-Type')) {
+        options.headers.set('Content-Type', 'application/vnd.api+json');
+    }
     return fetchJson(url, options);
 }